#2: Refactor files after merge of Google API branch
