--- my-app/server/.env
@@ -1,2 +1,0 @@
-GOOGLE_API_KEY=AIzaSyCjBy62P3tkGLfqy8HUUJ15f0pXPA1EdFA
-PORT=5000
